[Edit] 이미지 불러오기 관련 - DrinkImageViewModel 분리 #118
