+++ JUDA/Resource/Components/DrinkGridCell.swift
@@ -10,9 +10,7 @@ import Kingfisher
 
 // MARK: - 술 그리드 셀
 struct DrinkGridCell: View {
-    @EnvironmentObject private var drinkViewModel: DrinkViewModel
-    @State private var imageString: String = ""
-    @State private var isLoading: Bool = true
+    @StateObject private var drinkImageViewModel = DrinkImageViewModel()
     let drink: FBDrink
     
     // UITest - Drink 하트
@@ -31,8 +29,8 @@ struct DrinkGridCell: View {
             VStack(alignment: .leading, spacing: 10) {
                 // 술 사진
                 VStack(alignment: .center) {
-                    if isLoading {
-                        KFImage(URL(string: imageString))
+                    if drinkImageViewModel.isLoading {
+                        KFImage(URL(string: drinkImageViewModel.imageString))
                             .placeholder {
                                 CircularLoaderView(size: 20)
                                     .frame(width: 70, height: 103.48)
@@ -77,17 +75,9 @@ struct DrinkGridCell: View {
         .padding(10)
         // 이미지 불러오기
         .task {
-            if let imageName = drinkViewModel.getImageName(
-                                category: DrinkType(rawValue: drink.category) ?? DrinkType.all,
-                                detailedCategory: drink.type) {
-                if let imageString = await drinkViewModel.fetchImageUrl(imageName: imageName) {
-                    self.imageString = imageString
-                } else {
-                    self.isLoading = false
-                }
-            } else {
-                self.isLoading = false
-            }
+            await drinkImageViewModel
+                .getImageURLString(category: DrinkType(rawValue: drink.category) ?? DrinkType.all,
+                                   detailedCategory: drink.type)
         }
     }
